:suspect: refactor (tree.js, treePage.js, style.scss, makeTree.js) refactor to ES6
