+++ app/js/components/treePage.js
@@ -1,5 +1,4 @@
-var React = require('react');
-var D3Tree = require('./tree.js');
+'use strict';
 
 var treeData = [{
   id: 1,
@@ -64,81 +63,24 @@ var treeData = [{
   ]
 }];
 
-// var treeData = [{
-//        name: "flare",
-//        children: [
-//         {
-//          name: "analytics",
-//          children: [
-//           {
-//            name: "cluster",
-//            children: [
-//             {name: "AgglomerativeCluster", size: 3938},
-//             {name: "CommunityStructure", size: 3812},
-//             {name: "HierarchicalCluster", size: 6714},
-//             {name: "MergeEdge", size: 743}
-//            ]
-//           },
-//           {
-//            name: "graph",
-//            children: [
-//             {name: "BetweennessCentrality", "size": 3534},
-//             {name: "LinkDistance", size: 5731},
-//             {name: "MaxFlowMinCut", size: 7840},
-//             {name: "ShortestPaths", size: 5914},
-//             {name: "SpanningTree", size: 3416}
-//            ]
-//           },
-//           {
-//            name: "optimization",
-//            children: [
-//             {name: "AspectRatioBanker", size: 7074}
-//            ]
-//           }
-//          ]
-//         },
-//         {
-//          name: "animate",
-//          children: [
-//           {name: "Easing", size: 17010},
-//           {name: "FunctionSequence", size: 5842},
-//           {
-//            name: "interpolate",
-//            children: [
-//             {name: "ArrayInterpolator", size: 1983},
-//             {name: "ColorInterpolator", size: 2047},
-//             {name: "DateInterpolator", size: 1375},
-//             {name: "Interpolator", size: 8746},
-//             {name: "MatrixInterpolator", size: 2202},
-//             {name: "NumberInterpolator", size: 1382},
-//             {name: "ObjectInterpolator", size: 1629},
-//             {name: "PointInterpolator", size: 1675},
-//             {name: "RectangleInterpolator", size: 2042}
-//            ]
-//           },
-//           {name: "ISchedulable", size: 1041},
-//           {name: "Parallel", size: 5176},
-//           {name: "Pause", size: 449},
-//           {name: "Scheduler", size: 5593},
-//           {name: "Sequence", size: 5534},
-//           {name: "Transition", size: 9201},
-//           {name: "Transitioner", size: 19975},
-//           {name: "TransitionEvent", size: 1116},
-//           {name: "Tween", size: 6006}
-//          ]
-//         },
-//         {name: "data", size: 7000}
-//        ]
-//       }];
+import React from 'react';
+import D3Tree from './tree.js';
 
-var Page = React.createClass({
-  render: function() {
-    return (
-      <div>
-        <D3Tree treeData={treeData} />
-      </div>
-    );
+class Page extends React.Component {
+  constructor() {
+    super();
+    this.state = {
+      treeData: treeData
+    };
   }
-});
 
-module.exports = Page;
+  render() {
+    return (
+        <div className="treePage">
+          <D3Tree treeData={treeData} />
+        </div>
+      );
+  }
+}
+
+export default Page;
